obviously broken
git-svn-id: svn://svn.coreboot.org/coreboot/trunk@1820 2b7e53f0-3cfb-0310-b3e9-8179ed1497e1

+++ src/arch/ppc/boot/linuxbios_table.h
@@ -17,7 +17,7 @@ struct lb_record *lb_next_record(struct lb_record *rec);
 struct lb_record *lb_new_record(struct lb_header *header);
 struct lb_memory *lb_memory(struct lb_header *header);
 void lb_memory_range(struct lb_memory *mem, 
-	uint32_t type, unsigned long startk, unsigned long sizek);
+	uint64_t type, uint64_t startk, uint64_t sizek);
 struct lb_mainboard *lb_mainboard(struct lb_header *header);
 unsigned long lb_table_fini(struct lb_header *header);
